feat: Add user auth control functions

+++ server/controllers/userController.js
@@ -5,7 +5,14 @@ const { updateUserById, deleteUser, getUserInfo } = userService;
 
 const UserInfo = async (req, res) => {
 	const data = req.body
+	const user = req.user
+	
     try {
+		
+		if(!(user._id === data._id 
+			|| user.username === data.username 
+			|| user.email === data.email)) return res.status(403).json({status: false, message: "Yetkiniz yok"})
+		
 		if(!data) return res.status(204).json({status: false, message: "İstek boş"})
 		
 		const result = await getUserInfo(data)
@@ -26,7 +33,10 @@ const UserInfo = async (req, res) => {
 const UpdateUser = async (req, res) => {
 	const data = req.body
 	const id = req.params.id
+	
     try {
+		if(!(req?.user?._id === id)) return res.status(403).json({status: false, message: "Yetkiniz yok"})
+		
 		if(!data) return res.status(204).json({status: false, message: "İstek boş"})
 		
 		const result = await updateUserById({id, ...data})
@@ -48,7 +58,10 @@ const UpdateUser = async (req, res) => {
 
 const DeleteToUser = async (req, res) => {
 	const id = req.params.id
+	
     try {
+		if(!(req?.user?._id === id)) return res.status(403).json({status: false, message: "Yetkiniz yok"})
+
 		if(!id) return res.status(204).json({status: false, message: "Boş kimlik"})
 		
 		const result = await deleteUser(id)
